agregando el metodo de guardado de archivos

+++ src/main/java/com/srgi/dto/UExternoDTO.java
@@ -7,7 +7,7 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UExternoDTO extends UsuarioDTO{
-    private int cuil;
+    private long cuil;
     private String descripcion;
     private String empresa;
 }
